integration: move templates to nwo/template
This change moves the template files into a separate package and updates
the name to reduce the stutter.
