+++ integration/nwo/template/core_template.go
@@ -4,9 +4,9 @@ Copyright IBM Corp. All Rights Reserved.
 SPDX-License-Identifier: Apache-2.0
 */
 
-package nwo
+package template
 
-const DefaultCoreTemplate = `---
+const DefaultCore = `---
 logging:
   format: '%{color}%{time:2006-01-02 15:04:05.000 MST} [%{module}] %{shortfunc} -> %{level:.4s} %{id:03x}%{color:reset} %{message}'
