+++ integration/nwo/templates.go
@@ -6,6 +6,8 @@ SPDX-License-Identifier: Apache-2.0
 
 package nwo
 
+import "github.com/hyperledger/fabric/integration/nwo/template"
+
 // Templates can be used to provide custom templates to GenerateConfigTree.
 type Templates struct {
 	ConfigTx string `yaml:"configtx,omitempty"`
@@ -18,26 +20,26 @@ func (t *Templates) ConfigTxTemplate() string {
 	if t.ConfigTx != "" {
 		return t.ConfigTx
 	}
-	return DefaultConfigTxTemplate
+	return template.DefaultConfigTx
 }
 
 func (t *Templates) CoreTemplate() string {
 	if t.Core != "" {
 		return t.Core
 	}
-	return DefaultCoreTemplate
+	return template.DefaultCore
 }
 
 func (t *Templates) CryptoTemplate() string {
 	if t.Crypto != "" {
 		return t.Crypto
 	}
-	return DefaultCryptoTemplate
+	return template.DefaultCrypto
 }
 
 func (t *Templates) OrdererTemplate() string {
 	if t.Orderer != "" {
 		return t.Orderer
 	}
-	return DefaultOrdererTemplate
+	return template.DefaultOrderer
 }
